+++ src/components/badge/badge.tsx
@@ -5,7 +5,6 @@ import { BadgeColor, BadgeShape, BadgeSize } from './badge-interface';
  * The badge component is a small label, generally appearing inside or in proximity to another larger interface component,
  * representing a status, property, or some other metadata.  
  *
- * @cssprop --wcs-badge-lighter-percentage - Define the opacity rate to apply to the badge
  * @cssprop --wcs-badge-font-weight - Define the font weight of the badge
  * @cssprop --wcs-badge-height-l - Large height of the badge
  * @cssprop --wcs-badge-font-size-l - Large font size of the badge
@@ -17,10 +16,14 @@ import { BadgeColor, BadgeShape, BadgeSize } from './badge-interface';
  * @cssprop --wcs-badge-border-radius-circular - Circular border radius of the badge
  * 
  * @cssprop --wcs-badge-background-color - Background color of the badge
+ * @cssprop --wcs-badge-background-color-lighter - Background color of the lighter badge
  * @cssprop --wcs-badge-color - Text color of the badge
+ * @cssprop --wcs-badge-color-lighter - Text color of the lighter badge
  * @cssporp --wcs-badge-gap - The gap inside the badge
  * 
- * @cssprop --wcs-badge-padding-horizontal - Horizontal padding of the badge
+ * @cssprop --wcs-badge-padding-horizontal-l - Horizontal padding of the size L badge
+ * @cssprop --wcs-badge-padding-horizontal-m - Horizontal padding of the size M badge
+ * @cssprop --wcs-badge-padding-horizontal-s - Horizontal padding of the size S badge
  */
 @Component({
     tag: 'wcs-badge',